+++ src/commands/mod.rs
@@ -1,4 +1,5 @@
 pub mod token;
+pub mod oauthv2;
 pub mod esignature;
 pub mod ai;
 pub mod sms;