+++ tests/commands-test/catasto.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+cargo run -q -- -S catasto territorio
+cargo run -q -- -S catasto requests
+cargo run -q -- -S catasto visura-catastale