+++ tests/commands-test/oauth.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+cargo run -q -- -S oauth list
+cargo run -q -- -S oauth scopes
+cargo run -q -- -S oauth credit