+++ tests/commands-test/oauthv2.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+cargo run -q -- -S oauthv2 scopes list
+cargo run -q -- -S oauthv2 tokens list
+cargo run -q -- -S oauthv2 wallet info