+++ tests/commands-test/pec.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+cargo run -q -- -S pec list
+cargo run -q -- -S pec comunica-pec