+++ tests/commands-test/realestate.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+cargo run -q -- -S realestate rmv --address "Via Roma 1" --municipality "Roma"