+++ tests/commands-test/smsv2.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+cargo run -q -- -S smsv2 list