+++ tests/commands-test/ufficiopostale.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+cargo run -q -- -S ufficiopostale pricing
+cargo run -q -- -S ufficiopostale nazioni
+cargo run -q -- -S ufficiopostale raccomandate
+cargo run -q -- -S ufficiopostale telegrammi