+++ tests/commands-test/visurecamerali.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+cargo run -q -- -S visurecamerali ordinaria-impresa-individuale
+cargo run -q -- -S visurecamerali ordinaria-societa-capitale
+cargo run -q -- -S visurecamerali soci-attivi